fix(windows): CIのビルド失敗3件を修正し、WinAppSDKヘッダーの取り込み方式を変更する
CI(windows-2025)で初めてビルドして判明した3件を修正する。**CI を最初の
検証として使った結果であり、推測ではない。**

## 1. C4245 が警告エラー扱いで C2220 になっていた

```
audio_conversion.cpp(153,35): error C2220: the following warning is treated
  as an error
audio_conversion.cpp(153,35): warning C4245: 'argument': conversion from
  '__MIDL___MIDL_itf_mfreadwrite_0000_0001_0001' to 'DWORD',
  signed/unsigned mismatch
```

`MF_SOURCE_READER_ALL_STREAMS` / `MF_SOURCE_READER_FIRST_AUDIO_STREAM` は
符号付きの無名 enum であり、`IMFSourceReader` の各メソッドが取るのは
`DWORD`(符号なし)である。4箇所に `static_cast<DWORD>` を付けた。

## 2. `<experimental/coroutine>` の非推奨エラー

```
experimental/coroutine(37,1): error C2338: static assertion failed:
  'error STL1011: ... deprecated by Microsoft and will be REMOVED SOON'
```

Flutter の既定(C++17)では `winrt/base.h` が非推奨の
`<experimental/coroutine>` へ落ちる。WinRT 実装のビルド時に C++20 を要求する
ようにした。`_SILENCE_EXPERIMENTAL_COROUTINE_DEPRECATION_WARNINGS` は
「SOON に削除される」と明言されているものを延命するだけなので使わない。

## 3. WinAppSDK のプロジェクションヘッダーが取り込めていなかった(本丸)

```
winrt_includes.h(37,10): error C1083: Cannot open include file:
  'winrt/Microsoft.Windows.AI.h': No such file or directory
```

`winrt/base.h` や `winrt/Windows.Foundation.h`(Windows SDK 同梱)は開けて
いるため C++/WinRT 自体は使える。開けないのは WinAppSDK のプロジェクション
ヘッダーであり、これらは Windows SDK に含まれず NuGet パッケージ内の .winmd
から `cppwinrt.exe` が生成する。**CMake が生成する .vcxproj へ
`VS_PACKAGE_REFERENCES` を足すだけでは復元も生成も走らなかった。**

Microsoft は Flutter 向けの公式手順を用意している:
https://learn.microsoft.com/en-us/windows/apps/dev-tools/winapp-cli/guides/flutter
アプリ側で `winapp init` を実行すると WinAppSDK のヘッダー一式が
`<app>/.winapp/include` に展開され、それを include path に足して
`WindowsApp.lib` をリンクすればよい。

**この手順はアプリ側でどのみち必要である。** Windows AI APIs はパッケージ
識別(MSIX + systemAIModels capability)が無いと動かず、その識別を与えるのが
winapp CLI だからである。プラグインのためだけに増える手順ではない。

そこで `VS_PACKAGE_REFERENCES` 方式をやめ、`windows/CMakeLists.txt` が
`.winapp/include` を自動検出する方式にした:

- 見つかった場合  → WinRT 実装をビルドする
- 見つからない場合 → 全 API 呼び出しが「`winapp init` を実行せよ」という
  明示的なエラーで失敗する実装をリンクする。黙って `unavailable` を返す
  ようなフォールバックはしない

どちらになったかは構成時に `message(STATUS ...)` で必ず出力する。
`OFFLINE_STT_WINDOWS_WINAPP_INCLUDE_DIR` で場所を明示することもできる。

## 正直に書いておくべきこと

CI ランナーは winapp CLI を持たないため、**Windows ジョブは WinRT 実装を
ビルドしない**。検証されるのは WinRT に触れない部分(Pigeon 受け口・
スレッド調停・エラー分類・Media Foundation 変換)のコンパイルだけである。
**WinRT 実装そのもののコンパイルは依然として一度も通っていない。**
Issue #58 の実機検証でしか確認できない。この旨を ci.yml・CMakeLists.txt・
packages/offline_stt_windows/README.md §9 に明記した。

+++ packages/offline_stt_windows/windows/audio_conversion.cpp
@@ -150,9 +150,17 @@ AudioConversionResult ConvertToPcmWav(const std::string& input_path_utf8) {
   }
 
   // 音声ストリームだけを有効にする。
-  hr = reader->SetStreamSelection(MF_SOURCE_READER_ALL_STREAMS, FALSE);
+  //
+  // `MF_SOURCE_READER_ALL_STREAMS` / `MF_SOURCE_READER_FIRST_AUDIO_STREAM` は
+  // `MF_SOURCE_READER_CONTROL_FLAG` と同じヘッダーで定義された**符号付き**の
+  // 無名 enum であり、`IMFSourceReader` の各メソッドが取るのは `DWORD`
+  // (符号なし)である。MSVC は Flutter の既定設定(警告をエラー扱い)の下で
+  // これを C4245「signed/unsigned mismatch」として拒否する。実際に CI
+  // (windows-2025 / MSVC 14.51)で `error C2220` となったため、明示的に
+  // `static_cast<DWORD>` を書いている。値自体は変わらない。
+  hr = reader->SetStreamSelection(static_cast<DWORD>(MF_SOURCE_READER_ALL_STREAMS), FALSE);
   if (SUCCEEDED(hr)) {
-    hr = reader->SetStreamSelection(MF_SOURCE_READER_FIRST_AUDIO_STREAM, TRUE);
+    hr = reader->SetStreamSelection(static_cast<DWORD>(MF_SOURCE_READER_FIRST_AUDIO_STREAM), TRUE);
   }
   if (FAILED(hr)) {
     result.error = DecodeFailed("SetStreamSelection", hr);
@@ -184,7 +192,7 @@ AudioConversionResult ConvertToPcmWav(const std::string& input_path_utf8) {
     result.error = DecodeFailed("MFCreateMediaType", hr);
     return result;
   }
-  hr = reader->SetCurrentMediaType(MF_SOURCE_READER_FIRST_AUDIO_STREAM, nullptr,
+  hr = reader->SetCurrentMediaType(static_cast<DWORD>(MF_SOURCE_READER_FIRST_AUDIO_STREAM), nullptr,
                                    output_type.Get());
   if (FAILED(hr)) {
     // 要求した 16kHz モノラル 16-bit へ変換できない場合(入力に音声ストリーム
@@ -233,7 +241,7 @@ AudioConversionResult ConvertToPcmWav(const std::string& input_path_utf8) {
   for (;;) {
     DWORD stream_flags = 0;
     ComPtr<IMFSample> sample;
-    hr = reader->ReadSample(MF_SOURCE_READER_FIRST_AUDIO_STREAM, 0, nullptr,
+    hr = reader->ReadSample(static_cast<DWORD>(MF_SOURCE_READER_FIRST_AUDIO_STREAM), 0, nullptr,
                             &stream_flags, nullptr, sample.GetAddressOf());
     if (FAILED(hr)) {
       result.error = DecodeFailed("ReadSample", hr);
